refactored hello world app naming
