+++ hello-world/pom.xml
@@ -9,9 +9,9 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.angel</groupId>
-	<artifactId>hello-world-java17</artifactId>
+	<artifactId>hello-world</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>hello-world-java17</name>
+	<name>hello-world</name>
 	<description>Demo project for Spring Boot</description>
 	<properties>
 		<java.version>17</java.version>
